Add -bang to skip jumping to first arg
:args accepts a bang to drop the current file (like :e!), so that's not
what we want. Instead, we want to clear the arg list (using the
arbitrary high number given by the documentation) and then add our
arguments. If asked to jump to the first, then jump.

I think it makes more sense for the bang to jump to the first arg, but
changing that would break other users.