Use keepjumps to avoid adding unnecessary jump
Using :args jumps to the first file. For some reason this adds two
entries to the jumplist. I don't quite understand, but I don't think we
want to mess with the jumplist so we can jump back to where we came
from.

Verified with:
    gvim -Nu NONE +'source sensible.vim' +'source qargs.vim' +'grep! dave *.cmd ' README.rst
    :Qargs
    <C-o>

Before this change, nothing happens.
After this change, you jump back to the README.